ingestion/phase-4-6-chroma/requirements.txt
chromadb>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
